Standard PSR-1 ajout class document,documentStore,calculator,calcAdvenced
