+++ serie3/exercice3-2.php
@@ -1,9 +1,9 @@
 <?php
+  session_start();
   $serie = '3';
   $exercice = '3-2';  
   include('../template/header.php');
-?>
-    
+?>  
     <div class="container">
         <div class="row">
             <div class="col-sm">
@@ -15,69 +15,11 @@
             </div>
         </div>
 <?php  
-    if(isset($_POST['valider']))
-    {
-?>
-    <br>
-    <h3>Merci</h3>
-    <p>Merci pour avoir rempli le formulaire d'inscription. Voici un résumé des informations saisies : </p> 
-    <table class="table table-striped">
-        <tbody>
-            <tr>
-                <th>Prénom</th>
-                <td><?= $_POST['first_name'] ?></td>
-            </tr> 
-            <tr>
-                <th>Nom</th>
-                <td><?= $_POST['last_name'] ?></td>
-            </tr> 
-            <tr>
-                <th>Mot de passe</th>
-                <td><?= $_POST['password'] ?></td>
-            </tr> 
-            <tr>
-                <th>Confirmation mot de passe</th>
-                <td><?= $_POST['confirm_password'] ?></td>
-            </tr> 
-            <tr>
-                <th>Genre</th>
-                <td><?= $_POST['genre'] ?></td>
-            </tr> 
-            <tr>
-                <th>Langage favori</th>
-                <td><?= $_POST['language'] ?></td>
-            </tr> 
-            <tr>
-                <th>Souhaitez-vous recevoir notre newsletter ?</th>
-                <td><?= $_POST['newsletter'] ?></td>
-            </tr>
-            <tr>
-                <th>Commentaires</th>
-<?php
-        if(!empty($_POST['coments']))
+        if ($_GET['etape'] == 1 || !isset($_GET['etape']))
         {
-?>
-                <td><?= $_POST['comments'] ?></td>
-            </tr> 
-<?php   } 
-        else
-        {
-?>
-                <td>No comments</td>
-            </tr>
-
-<?php   } ?>
-
-        </tbody>
-    </table>
-
-<?php  
-    }
-    else
-    {
-?>        
+?>      
         <h3>Inscription : Etape 1</h3><br>
-        <form action="" method="post">
+        <form action="exercice3-3.php?etape=2" method="post">
             <div class="row">
                 <div class="col">
                     <label>Prénom</label>
@@ -96,6 +38,23 @@
                 <label for="exampleInputPassword1">Confirmation mot de passe</label>
                 <input type="password" class="form-control" name="confirm_password" placeholder="Password" required>
             </div>
+            <br>
+            <button type="submit" name="suivant" class="btn btn-primary">Suivant</button>
+        </form>
+<?php              
+        }
+        elseif ($_GET['etape'] == 2)
+        {
+            if (!isset($_SESSION['first_name'])) 
+            {
+                $_SESSION['first_name'] = $_POST['first_name'];
+                $_SESSION['last_name'] = $_POST['last_name'];
+                $_SESSION['password'] = $_POST['password'];
+                $_SESSION['confirm_password'] = $_POST['confirm_password'];
+            }
+?>
+        <h3>Inscription : Etape 2</h3><br>
+        <form action="exercice3-3.php?etape=3" method="post">
             <div class="form-check form-check-inline">
                 <input class="form-check-input" type="radio" name="genre" id="inlineRadio1" value="homme" required> 
                 <label class="form-check-label" for="inlineRadio1">Homme</label>
@@ -116,11 +75,27 @@
                 </select>
                 </div>
             </div>
+            <br>
+            <a href="exercice3-3.php?etape=1"><button type="button" name="retour" class="btn btn-primary">Retour</button></a>
+            <button type="submit" name="suivant" class="btn btn-primary">Suivant</button>
+        </form>
+<?php
+        }
+        elseif ($_GET['etape'] == 3)
+        {
+            if (!isset($_SESSION_['genre'])) 
+            {
+                $_SESSION['genre'] = $_POST['genre'];
+                $_SESSION['language'] = $_POST['language'];
+            }    
+?>
+        <h3>Inscription : Etape 3</h3><br>
+        <form action="exercice3-3.php?etape=fini" method="post">
             <div class="form-group row">
                 <label class="col-sm-2">Souhaitez-vous recevoir notre newsletter ?</label>
                 <div class="col-sm-10">
                     <div class="form-check">
-                        <input type="checkbox" class="form-check-input" name="newsletter" id="gridCheck1" required>
+                        <input type="checkbox" class="form-check-input" name="newsletter" id="gridCheck1" value="yes">
                     </div>
                 </div>
             </div>
@@ -129,11 +104,79 @@
                 <textarea class="form-control" name="comments" rows="3"></textarea>
             </div>
             <br>
-            <button type="reset" class="btn btn-primary">Rénitialiser</button>
+            <a href="exercice3-3.php?etape=2"><button type="button" name="retour" class="btn btn-primary">Retour</button></a>
             <button type="submit" name="valider" class="btn btn-primary">Envoyer</button>
         </form>
 <?php  
-    }
+        }
+        elseif ($_GET['etape'] == "fini") 
+        {
+?>
+                <br>
+    <h3>Merci</h3>
+    <p>Merci pour avoir rempli le formulaire d'inscription. Voici un résumé des informations saisies : </p> 
+    <table class="table table-striped">
+        <tbody>
+            <tr>
+                <th>Prénom</th>
+                <td><?= $_SESSION['first_name'] ?></td>
+            </tr> 
+            <tr>
+                <th>Nom</th>
+                <td><?= $_SESSION['last_name'] ?></td>
+            </tr> 
+            <tr>
+                <th>Mot de passe</th>
+                <td><?= $_SESSION['password'] ?></td>
+            </tr> 
+            <tr>
+                <th>Confirmation mot de passe</th>
+                <td><?= $_SESSION['confirm_password'] ?></td>
+            </tr> 
+            <tr>
+                <th>Genre</th>
+                <td><?= $_SESSION['genre'] ?></td>
+            </tr> 
+            <tr>
+                <th>Langage favori</th>
+                <td><?= $_SESSION['language'] ?></td>
+            </tr> 
+            <tr>
+                <th>Souhaitez-vous recevoir notre newsletter ?</th>
+<?php
+        if(!empty($_POST['newsletter']))
+        {
+            echo "<td>." . $_POST['newsletter'] . "</td> 
+            </tr>";  
+        } 
+        else
+        {
+            echo "
+                <td>no</td>
+            </tr>";
+        } 
+            echo "
+            <tr>
+                <th>Commentaires</th>";
+        if (!empty($_POST['comments']))
+        {
+            echo "
+                <td>" . $_POST['comments'] . "</td>
+            </tr>";   
+        } 
+        else
+        {
+            echo "
+                <td>No comments</td>
+            </tr>";
+        } 
+        session_destroy();
+?>
+        </tbody>
+    </table>
+
+<?php     
+        }
 ?>
     </div>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/js/bootstrap.bundle.min.js" integrity="sha384-ygbV9kiqUc6oa4msXn9868pTtWMgiQaeYH7/t7LECLbyPA2x65Kgf80OJFdroafW" crossorigin="anonymous"></script>
